Add initial theme option to the app config
I'm not quite sure what options you want to expose in the
application config for flash message, so I've currently just
exposed a flashTheme config option. This is what the alertType
will use to infer the class prefix. One can still explicitly
set the messageType in the component declaratively, I've added
a test to verify that as well.

Related to #10