allow passing arbitrary options to flash messages
Sometimes an application wants to customize how its flashes are
rendered on a per-flash basis. This allows the application to set
arbitrary options onto the instance of `FlashMessage` so that display
can be customized.

For example, in the OSX update message, there is a non-trivial UI
contained entirely inside the flash message, complete with buttons and
dropdowns.

By allowing the caller of `flashes.add()` to pass any context data along
to the flash object, the template that ultimately renders the flash can
be as rich as it needs to be.