Add id to calls to Ember.warn(). Return old value in read only CP set
Passing id to Ember.warn avoids deprecation messages from Ember.

When using the new compute property syntax, the return value from the
"set" function is cached as the new value of the CP. The next time that CP
is required through `Ember.get()` or `emberObject.get()`, the cached value is
returned and the "get" function specified in the computed property is *not*
called.

Returning this inside those functions made those CPs take on the value of the
object itself.