Work around issue when using `component` reserved word.
In Ember's alpha channel, naming a block param `component` results in a
very difficult to parse error message.  This changes the tests and demo
to use `comp` instead of `component`. I have also opened an issue
upstream for this, so that we can either get it fixed or have a much
nicer assertion/error when using reserved words.