provide default config/overrides to FlashMessageService w/o injections (#284)
While I was writing a component integraiton test in my app, I noticed
that I couldn't call `this.flashMessages.success` or any of the built in
convenience functions that target the bootstrap notifications.

This seems to be because, unless an application test runs first, the
service isn't getting set up with the right config. It looks like the
service is expecting it from only the initializer right now.

I changed app/services/flash-message.js to import the config from the
application, so that it is always set up correctly regardless of the
test context.

I also extracted the logic for merging the addon defaults and the app
overrides into a utility module so we could use the same logic in both
places as much as possible.